+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade Atata package to v3.0.0.
+- Change `HtmlValidationOptions.HtmlValidatePackageVersion` property default value to `"8.18.1"`.
 
 ### Removed