+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade Atata package reference to v3.4.0.
+- Change `HtmlValidationOptions.HtmlValidatePackageVersion` property default value to `"8.29.0"`.
 
 ## [3.1.0] - 2024-05-18